+++ ChangeLog.md
@@ -29,6 +29,7 @@
     - Define new signupSubmit acUserLogin configuration parameter
     - Modal options passed to acUserLogin are also passed to modal rendering component
     - Fix and improve signup panel stylesheet
+    - Add errFn() and successFn callback function to createUser() options
 
 ### 1.4.0